Replace wisper by wisper_kafka
Events get published to a Kafka topic, but the Kafka message has an
empty `subscriber` section (actually an empty hash, `{}`).  This
confuses the Kafka consumer, which expects it to be a string instead.

I suspect this is because I use a global subscriber, attached using
`Wisper.subscribe`, instead of subscribing to the producer directly.  In
this case, the subscriber is a controller, so I cannot really subscribe
to it ahead of time.